+++ NEWS.md
@@ -1,5 +1,14 @@
 
 
+# vcfR 1.0.0.9000
+
+* Added tidyr compatibility.
+* `write.vcf()` now uses mask = TRUE.
+* `maf()` provides counts and frequency for the minor (or other) allele.
+* `create.chromR()` now handles instances with no seq and the annotation position exceeds the greatest VCF POS.
+* `read.vcfR()` now handles tilde expansion.
+* `addID()` populates the non-missing values in the ID column of VCF data by concatenating the chromosome and position. 
+
 # vcfR 1.0.0
 Released on CRAN 2016-02-22.